Added initial imGUI implementation + Testing
(currently broken) imGUI has been added to the repo and CMake projects. The only issue now is that an imGUI header in the backend folder can't find the GLFW headers so currently getting compile-time errors. A potential fix to this would be to turn the imGUI backend folder into its own library and maybe give it GLFW as a target link library.